Split procstat_vm.c
